Add `pyproject.toml` file and AutoPub support
Contributors will henceforth be asked to include a `RELEASE.md` file
containing the release type (major, minor, patch) and a summary of the
changes in the pull request. This information will be used by the
continuous integration (CI) system to automatically publish a new
release to PyPI when the pull request is merged.

For more details, see: https://github.com/autopub/autopub