Fix transliteration and add types
Transliteration was still in python2 mode and never upgraded. Since
this module should no longer support python 2.x rewrite it to python 3.

Python 3's join() doesn't allow you to combine bytes and str, so we
have to encode the delimiter as well.

Last but not least, document how translitcodec works, so we don't have
to find this out the hard way.

Also type this module.
Utils are only used internally, but type them anyway as they can be
referenced trough dotted notation in settings or people may import and
use some of the functions here.